chore: cleanup & move engine
